Moved user signup endpoint to user's group
